feat: delete view에 사용자와 글의 작성자가 같을때만 삭제가 가능하게 수정

+++ 03_02_db_many_to_one_relationships/articles/views.py
@@ -46,7 +46,8 @@ def detail(request, article_id):
 @login_required
 def delete(request, article_id):
     article = Article.objects.get(pk=article_id)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
